+++ src/apps/companies/apps/edit-one-list/client/EditOneListForm.jsx
@@ -7,7 +7,7 @@ import {
   Main,
   FieldRadios,
   Step,
-  AdviserTypeAhead,
+  FieldAdvisersTypeahead,
 } from '../../../../../client/components'
 import Form from '../../../../../client/components/Form'
 import { TASK_SAVE_ONE_LIST_DETAILS } from './state'
@@ -62,12 +62,12 @@ function EditOneListForm({
 
             {values.one_list_tier !== NONE && (
               <Step name="oneListAdvisers">
-                <AdviserTypeAhead
+                <FieldAdvisersTypeahead
                   name={ACCOUNT_MANAGER_FIELD_NAME}
                   label="Global Account Manager"
                   required="Select at least one adviser"
                 />
-                <AdviserTypeAhead
+                <FieldAdvisersTypeahead
                   name={ONE_LIST_TEAM_FIELD_NAME}
                   label="Advisers on the core team (optional)"
                   isMulti={true}
