+++ src/client/components/index.jsx
@@ -24,7 +24,7 @@ export { default as StatusMessage } from './StatusMessage'
 export { default as SummaryList } from './SummaryList'
 export { default as SummaryTable } from './SummaryTable'
 export { default as Tag } from './Tag'
-export { default as AdviserTypeAhead } from './AdviserTypeAhead'
+export { default as FieldAdvisersTypeahead } from './Form/elements/FieldAdvisersTypeahead'
 export { default as Typeahead } from './Typeahead/Typeahead'
 export { default as DropdownMenu } from './DropdownMenu'
 export { default as FieldWrapper } from './Form/elements/FieldWrapper/'
